[feature/add-exception-listener] Transform listerner to subscriber to manage the lexer used
