--- Resources/config/services/listener.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <parameters>
-        <parameter key="front_render_bundle.twig_listener.class">Chris\Bundle\FrontRenderBundle\Listener\TwigListener</parameter>
-        <parameter key="front_render_bundle.exception_listener.class">Chris\Bundle\FrontRenderBundle\Listener\ExceptionListener</parameter>
-    </parameters>
-
-    <services>
-        <service id="front_render_bundle.twig_listener" class="%front_render_bundle.twig_listener.class%">
-            <argument type="service" id="twig" />
-            <tag name="kernel.event_listener" event="kernel.request" method="updateTagTwig"/>
-        </service>
-        <service id="front_render_bundle.exception_listener" class="%front_render_bundle.exception_listener.class%">
-            <argument type="service" id="twig" />
-            <tag name="kernel.event_listener" event="kernel.exception" method="onKernelException"/>
-        </service>
-    </services>
-</container>
