+++ Resources/config/services/subscriber.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+
+    <parameters>
+        <parameter key="front_render_bundle.render_subscriber.class">Chris\Bundle\FrontRenderBundle\Subscriber\RenderSubscriber</parameter>
+    </parameters>
+
+    <services>
+        <service id="front_render_bundle.render_subscriber" class="%front_render_bundle.render_subscriber.class%">
+            <argument type="service" id="twig" />
+            <tag name="doctrine.event_subscriber" />
+        </service>
+    </services>
+</container>
